librispeech_loader: add out memory reading from h5py. Small refactoring of transform function in esc reader
